+++ src/main/java/module-info.java
@@ -18,9 +18,9 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-module gradle.migration.data.main
+module io.github.astrapisixtynine.gradle.migration.data
 {
-	requires lombok;
+	requires static lombok;
 	requires org.apache.commons.lang3;
 	requires org.apache.commons.text;
 	requires silly.collection;